Renames nnue_ to nnue
