+++ src/board/position/mod.rs
@@ -3,11 +3,11 @@ use std::ops::Deref;
 use crate::bit_move::MoveType;
 use crate::constants::params::PIECE_VALUES;
 use crate::constants::{BLACK_KING_CASTLING_MASK, BLACK_QUEEN_CASTLING_MASK, WHITE_KING_CASTLING_MASK, WHITE_QUEEN_CASTLING_MASK};
-use crate::nnue_::accumulator::Feature;
+use crate::nnue::accumulator::Feature;
 use crate::{bit_move::Move, move_scope::MoveScope, squares::Square};
-use crate::nnue_::network::NNUEState;
+use crate::nnue::network::NNUEState;
 use crate::color::Color::{self, *};
-use crate::nnue_::constants::customKA0::*;
+use crate::nnue::constants::customKA0::*;
 use crate::squares::Square::*;
 use super::castling::Castling;
 use super::{piece::{Piece, Piece::*}, state::board::Board};
